requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohttp=3.7.4.post0=py310he2143c4_1
async-timeout=3.0.1=py_1000
attrs=21.4.0=pyhd8ed1ab_0
bzip2=1.0.8=h3422bc3_4
ca-certificates=2022.6.15=h4653dfc_0
chardet=4.0.0=py310hbe9552e_3
click=8.1.3=py310hbe9552e_0
discord-pretty-help=1.3.4=pypi_0
discord.py=1.7.3=pyhd8ed1ab_0
dnspython=2.2.1=pypi_0
idna=3.3=pyhd8ed1ab_0
libcxx=14.0.6=h04bba0f_0
libffi=3.4.2=h3422bc3_5
libzlib=1.2.12=ha287fd2_1
multidict=6.0.2=py310hf8d0d8f_1
ncurses=6.3=h07bb92c_1
openssl=3.0.4=ha287fd2_2
pip=22.1.2=pyhd8ed1ab_0
pymongo=4.1.1=py310h1105856_0
python=3.10.5=h4eee789_0_cpython
python-dotenv=0.20.0=pyhd8ed1ab_0
python_abi=3.10=2_cp310
readline=8.1.2=h46ed386_0
setuptools=62.6.0=py310hbe9552e_0
sqlite=3.39.0=h40dfcc0_0
tk=8.6.12=he1e0b03_0
typing-extensions=4.2.0=hd8ed1ab_1
typing_extensions=4.2.0=pyha770c72_1
tzdata=2022a=h191b570_0
wheel=0.37.1=pyhd8ed1ab_0
xz=5.2.5=h642e427_1
yarl=1.7.2=py310hf8d0d8f_2
zlib=1.2.12=ha287fd2_1
